Refactor: separate ExamService into interface and implementation
Refactor: separate ExamService into interface and implementation
